requirements.txt
dj-database-url==0.3.0
django-crispy-forms==1.5.1
Django==1.8.2
gunicorn==19.3.0
psycopg2==2.6
SQLAlchemy==1.0.8
whitenoise==2.0.3
django-postgrespool==0.3.0
django-haystack==2.4.0
whoosh==2.7.0
django-registration-redux==1.2
